Refactor internalRenderMarkdownInline
