--- packages/component/src/hooks/internal/useInternalRenderMarkdownInline.js
@@ -1,9 +1,0 @@
-import useWebChatUIContext from './useWebChatUIContext';
-
-function useInternalRenderMarkdownInline() {
-  const { internalRenderMarkdownInline } = useWebChatUIContext();
-
-  return internalRenderMarkdownInline;
-}
-
-export default useInternalRenderMarkdownInline;
